INITIAL COMMIT: webcrawler and edge-generator
This is the intial commit for webcrawler. Right now
it is just a badly written script that's getting a
job done.

The edge-crawler script produces a Gephi-compatible
edge list for creating network graphs. The file is
also useful for deriving 404 and other such lists.

Using the Issue Tracker to track issues and features.